+++ plivo/resources/numbers.py
@@ -35,8 +35,8 @@ class Numbers(PlivoResourceInterface):
         app_id=[optional(of_type(six.text_type))],
         verification_info=[optional(of_type_exact(dict))],
         cnam_lookup=[optional(of_type(six.text_type))],
-        ha_enabled=[optional(of_type(bool))],)
-    def buy(self, number, app_id=None, verification_info=None, cnam_lookup=None, ha_enabled=None):
+        ha_enable=[optional(of_type(bool))],)
+    def buy(self, number, app_id=None, verification_info=None, cnam_lookup=None, ha_enable=None):
         return self.client.request('POST', ('PhoneNumber', number),
                                    to_param_dict(self.buy, locals()))