+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Change Log
+## [4.59.5](https://github.com/plivo/plivo-python/tree/v4.59.5) (2026-01-21)
+**Feature - HA Number support**
+- Added `ha_enable` optional parameter to PhoneNumber `buy` method for enabling High Availability Number functionality
+- Added `fallback_number` field in PhoneNumber buy response
+
 ## [4.59.4](https://github.com/plivo/plivo-python/tree/v4.59.4) (2026-01-13)
 **Feature - Profile API business_contact_email support**
 - Added `business_contact_email` parameter support to Profile `create` and `update` methods for PUBLIC entity types